get sparc64 almost building (fails due to my binutils at the moment)
git-svn-id: svn://coreboot.org/openbios/openbios-devel@60 f158a5a8-5612-0410-a976-696ce0be7e32

+++ include/sparc64/types.h
@@ -16,8 +16,8 @@
 /* cell based types */
 typedef int64_t		 cell;
 typedef uint64_t	ucell;
-typedef long long	dcell;
-typedef unsigned long long ducell;
+typedef __int128_t	dcell;
+typedef __uint128_t    ducell;
 
 #define bitspercell	(sizeof(cell)<<3)
 #define bitsperdcell	(sizeof(dcell)<<3)
